+++ src/stores/consumption/actions/laptimeWouldBeStintDurationMinutes.js
