+++ src/stores/consumption/configureStore.js
