+++ src/stores/consumption/constants/index.js
