+++ src/stores/consumption/reducers/laptimeReducer.js
