+++ src/stores/consumption/reducers/stintDetailsInitialState.js
