+++ src/stores/consumption/selectors/computeStintDurationValue.js
