+++ src/stores/consumption/selectors/formating.js
