+++ src/stores/consumption/selectors/lapConsumptionSelector.js
